woring on transfer status

+++ src/CoreCom/Source/WallTec.CoreCom.Sheard/CoreComInternalSignatures.cs
@@ -10,7 +10,7 @@ namespace WallTec.CoreCom.Sheard
         public const string CoreComInternal_ConnectInstallId = "CoreComInternal_ConnectInstallId";
         public const string CoreComInternal_PullCue = "CoreComInternal_PullCue";
     }
-    public enum TransferStatus : byte
+    public enum TransferStatus : int
     {
         New = 0,
         InProcess=1,
